[Refactor] 파일 분리
